Refactoring: FileClass implementation replaced to infrastructure
