+++ oetube/aspnet-core/OeTube/Infrastructure/FileClasses/VideoFiles/FramesDirectoryFileClass.cs
@@ -1,4 +1,4 @@
-﻿namespace OeTube.Domain.Infrastructure.Videos.VideoFiles
+﻿namespace OeTube.Infrastructure.FileClasses.VideoFiles
 {
 
     public abstract class FramesDirectoryFileClass : VideoFileClass
