+++ oetube/aspnet-core/OeTube/Infrastructure/FileClasses/VideoFiles/HlsSegmentFileClass.cs
@@ -1,6 +1,6 @@
 ﻿using OeTube.Domain.Entities.Videos;
 
-namespace OeTube.Domain.Infrastructure.Videos.VideoFiles
+namespace OeTube.Infrastructure.FileClasses.VideoFiles
 {
     public class HlsSegmentFileClass : ResolutionDirectoryFileClass
     {
