+++ oetube/aspnet-core/OeTube/Infrastructure/FileClasses/VideoFiles/SelectedFrameFileClass.cs
@@ -1,4 +1,4 @@
-﻿namespace OeTube.Domain.Infrastructure.Videos.VideoFiles
+﻿namespace OeTube.Infrastructure.FileClasses.VideoFiles
 {
     public class SelectedFrameFileClass : FramesDirectoryFileClass
     {
